LayoutTests/http/tests/security/contentSecurityPolicy/upgrade-insecure-requests/upgrade-insecure-audio-video-in-main-frame-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/resources/frame-with-insecure-audio-video.html requested insecure content from http://127.0.0.1:8443/resources/test.mp4. This content was automatically upgraded and should be served over HTTPS.

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/contentSecurityPolicy/upgrade-insecure-requests/resources/frame-with-insecure-audio-video.html requested insecure content from http://127.0.0.1:8443/resources/test.mp4. This content was automatically upgraded and should be served over HTTPS.

This test opens a window that loads insecure HTML5 audio and video. Although the main frame in the window is HTTPS, and it is asked to display insecure content, we should upgrade the insecure content to HTTPS and therefore not trigger a mixed content callback.
